Fixing Basic Functionality
I realized that I didn't actually have to commit for testing
(thankfully) and realized that my previous commit caused an infinite
loop, preventing interaction with the page. I fixed that and also made
the page work on a basic level.

I will probably continue to refine this page until it meets my
standards, but it is usable for now.