Adding JSON data files, and Script for making them
So I've been working on this for a while. I wanted to take the 3
plaintext files we were working from before and turn them into JSON
files which can be more easily turned into usable data without
additional parsing. This makes searching slightly faster (but it wasn't
really that slow to begin with) and will allow me to add features
without additional parsing.

As of this commit, the index file hasn't been changed to use the new
JSON files, so that's what I will work on next.