Changed the interface to actually use the JSON data. Also changed the Pokemon Sets data file to include name and number
I finally put in all the work to change the search() and pkSearch()
functions to actually use the JSON data. The pkSearch() function got
much simpler, and returns more compartmentalized data at a quicker
speed. Unfortuantely, I had to make the search() function (what actually
places everything in the HTML table so the user can see) more
complicated to make use of the data. The benefit is a more detailed
interface which should allow me to make other improvements down the road
more easily.

One entirely new feature is a trainer wildcard. Simply write * for the
trainer name, and you will get every set for the given Pokemon name.
This is useful if you forget to read the trainre name, or just want
every piece of data for that Pokemon.

Next order of business will probably be cleaning up all of the code,
since it's become rather messy now.