Adding functionality, separating JS and CSS from index, renaming json files, and expanding README
I added one major new piece of functionality: clicking on a result will
now highlight it in a light green. Clicking again removes the highlight.
This main use of this is to allow users to highlight a set once they've
confirmed which one the opponent is using, or to highlight multiple
possibilities as the opponent reveals more moves.

Another new feature is that the interface tells the user while
information is loading. This makes it much clearer to users with very
slow internet why the interface isn't doing anything. The message
disappears immediately when everything is loaded, and the user's query
(if they've already typed one) is automatically run.

More importantly, though, all CSS and almost all JavaScript was
separated out into multiple files which are now loaded via link and
script elements, respectively. This does increase loading times, but
should make the organization of things much clearer.

That said, the MaisonSearch.js file is still very long (longer now that
I've added lots of comments and a couple new functions), and so I'm
slightly urged to break it down further. However, doing so would cause a
further performance penalty (however small) and require more code to
make sure everything is loaded in-order.

The JSON files were renamed to remove the redundant "JSON", and to put
them in-line with the names used by the code.

Finally, I made a full write-up for the README, explaining how the
Battle Maison works, how to use the search engine, and how the search
engine works internally. A link to the page was also added at the top of
the README, and a link back to the repository was added to the page's
header, so users can easily go back-and-forth.

At this point the page is fully functional, so I'll probably merge this
into master. I'll start working on the Battle Tree (Sun & Moon) version
next, which shouldn't take too long as it will require minimum changes
outside of the data files.