Removing Extra Files and slight changes to index and readme
I removed JSONMaker.js and the old text files, which are no longer used
anywhere in the program. These can still be accessed by going back
through the Git history, but they won't be presented in the master
branch anymore.

I changed "people" in a comment on index.html to "users". Not important.

Lastly, I removed the reference to non-minified JS files in the
README.md file, since I decided not to minify anything. I don't want to
minify the JSON (since that would defeat the point of using a
human-readable format in the first place) and minifying the JS isn't
worth it, since the amount of data saved (before GZIP, of course) is
pidly. The files total 18 kb and I could get them down to maybe 10 kb
with minifying. Considering the JSON files are in the hundreds of kb, I
don't think it's worth losing the ability to read the code and comments
right in the browser's developer panel.