Removing unused style
There was a style rule in main.css for pre elements. This is a vestige
from my previous web tools which did make pre elements. In this one,
though, I never make or reference a pre element, so I just removed the
rule.

I also added a comment describing the background color of the body
element.