Adding Google Analytics
I'm adding analytics, mainly to check how many users currently use the
site, and how often. I may use this to make changes in the future. I may
also add advertisements in the future.

This change does not affect the use of the site in any way. The
information you type into any of the site's textboxes is not currently
recorded or transmitted. Only data about your use time, location, and
client (browser) will be trasmitted.